move code to detect commented JSON files to util.py
